+++ export_download/views/__init__.py
@@ -1,13 +1,17 @@
 from django.http import HttpResponseNotAllowed, Http404, HttpResponse
-from django.urls import reverse
-from django.views.generic.list import MultipleObjectMixin
+from django.views.generic.list import ListView
 
 from import_export.formats import base_formats
 from import_export.resources import Resource
 
 
-class ExportDownloadView(MultipleObjectMixin):
+class ResourceDownloadMixin:
     resource_class = None
+
+    resource_formats = ['csv', 'xls']
+
+    resource_class_parameter = 'resource_class'
+    resource_format_parameter = 'resource_format'
 
     _resource_format_map = {
         'csv': base_formats.CSV,
@@ -17,10 +21,7 @@ class ExportDownloadView(MultipleObjectMixin):
         'tsv': base_formats.TSV,
     }
 
-    resource_class_parameter = 'resource_class'
-    resource_format_parameter = 'resource_format'
-
-    resource_formats = ['csv', 'xls']
+    _download_parameter = 'download'
 
     def __init__(self, *args, **kwargs):
         """
@@ -28,6 +29,8 @@ class ExportDownloadView(MultipleObjectMixin):
         """
         super().__init__(*args, **kwargs)
 
+        if not issubclass(self.__class__, ListView):
+            raise NotImplementedError('You can use the ExportDownloadMixin only in a ListView')
         if not self._get_ressource_classes():
             raise NotImplementedError('Object {}.resource_class must be defined.'.format(self.__class__.__name__))
         for k in self._get_ressource_classes():
@@ -43,13 +46,6 @@ class ExportDownloadView(MultipleObjectMixin):
                 raise NotImplementedError('Format {} in {}.resource_class is not a valid '
                                           'resource_formats'.format(f, self.__name__))
 
-    def get_export_url(self, resource_class):
-        """
-        You can overwrite this method to return the download URL specified in your url.py
-        where calling .as_download()
-        """
-        return reverse('{}-export'.format(resource_class.Meta.model._meta.model_name))
-
     @classmethod
     def _get_ressource_classes(cls):
         """
@@ -62,11 +58,11 @@ class ExportDownloadView(MultipleObjectMixin):
         else:
             return [cls.resource_class]
 
-    def _get_resource_links(self, request):
+    def get_resource_links(self, request):
         """
         This method return a dict in the form:
         {
-            '<resource_format>': ['<download_link>', '<description>',
+            '<resource_format>': ['<download_link>', '<description>'],
             ...
         }
 
@@ -77,14 +73,13 @@ class ExportDownloadView(MultipleObjectMixin):
         for f in self.resource_formats:
             resource_links[f.lower()] = []
             for counter, resource_class in enumerate(self._get_ressource_classes()):
-                export_url = self.get_export_url(resource_class)
                 params = request.GET.copy()
                 params_class = {
                     self.resource_class_parameter: counter,
                     self.resource_format_parameter: f,
                 }
                 params.update(params_class)
-                link = export_url + "?" + self._to_url_params(params)
+                link = "?" + self._to_url_params(params)
                 # if there is a description field in the resource class
                 # we use it to display it as a description
                 description = getattr(resource_class, 'description', resource_class.__name__)
@@ -96,45 +91,42 @@ class ExportDownloadView(MultipleObjectMixin):
         """
         reeturn a kwarg in GET parameter format
         """
-        return "&".join('{}={}'.format(k, v) for k, v in d.items())
+        return 'download&' + "&".join('{}={}'.format(k, v) for k, v in d.items())
 
-    @classmethod
-    def as_download(cls, *args, **kwargs):
-        """
-        This method returns the view where the data is exported.
-        It is inspired by the .as_view() from django.
-        """
-        def view(request, *args, **kwargs):
-            if request.method != 'GET':
-                return HttpResponseNotAllowed(['GET'])
-            # We use the first resource class and first resource format
-            # as a default, when there are no parameters
-            resource_class = request.GET.get(cls.resource_class_parameter, 0)
-            resource_format = request.GET.get(cls.resource_format_parameter, cls._get_ressource_classes()[0])
+    def render_to_response(self, *args, **kwargs):
+        if self._download_parameter in self.request.GET:
+            return self.render_to_download_response(*args, **kwargs)
+        return super().render_to_response(*args, **kwargs)
 
-            if not resource_format:
-                raise Http404('You have to pass {} as GET parameter'.format(cls.resource_format_parameter))
+    def render_to_download_response(self, *args, **kwargs):
+        if self.request.method != 'GET':
+            return HttpResponseNotAllowed(['GET'])
+        # We use the first resource class and first resource format
+        # as a default, when there are no parameters
+        resource_class = self.request.GET.get(self.resource_class_parameter, 0)
+        resource_format = self.request.GET.get(self.resource_format_parameter, self._get_ressource_classes()[0])
 
-            selected_format = cls._resource_format_map.get(resource_format, None)
-            if not selected_format:
-                raise Http404('Export format {} not found'.format(resource_format))
+        if not resource_format:
+            raise Http404('You have to pass {} as GET parameter'.format(self.resource_format_parameter))
 
-            try:
-                resource_class_number = int(resource_class)
-            except:
-                raise Http404('Parameter {} must be an integer'.format(cls.resource_class_parameter))
-            if resource_class_number >= len(cls._get_ressource_classes()):
-                raise Http404('Parameter {}.{} does not exist'.format(cls.__name__, cls.resource_class_parameter))
+        selected_format = self._resource_format_map.get(resource_format, None)
+        if not selected_format:
+            raise Http404('Export format {} not found'.format(resource_format))
 
-            qs = cls.model.objects.all()
-            # If filer_class is defined try to filter against it.
-            # You need django-filter to use this feature.
-            if hasattr(cls, 'filter_class'):
-                if cls.filter_class:
-                    qs = cls.filter_class(request.GET, queryset=qs).qs
-            resource_class = cls._get_ressource_classes()[resource_class_number]
-            export = resource_class().export(qs)
-            r = getattr(export, selected_format.__name__.lower())
-            return HttpResponse(r, content_type=selected_format.CONTENT_TYPE)
+        try:
+            resource_class_number = int(resource_class)
+        except:
+            raise Http404('Parameter {} must be an integer'.format(self.resource_class_parameter))
+        if resource_class_number >= len(self._get_ressource_classes()):
+            raise Http404('Parameter {}.{} does not exist'.format(self.__class__.__name__, self.resource_class_parameter))
 
-        return view
+        qs = self.model.objects.all()
+        # If filer_class is defined try to filter against it.
+        # You need django-filter to use this feature.
+        if hasattr(self, 'filter_class'):
+            if self.filter_class:
+                qs = self.filter_class(self.request.GET, queryset=qs).qs
+        resource_class = self._get_ressource_classes()[resource_class_number]
+        export = resource_class().export(qs)
+        r = getattr(export, selected_format.__name__.lower())
+        return HttpResponse(r, content_type=selected_format.CONTENT_TYPE)